+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/PodTemplateStep.java
@@ -35,6 +35,7 @@ public class PodTemplateStep extends Step implements Serializable {
     private final String label;
     private final String name;
 
+    private String namespace;
     private List<ContainerTemplate> containers = new ArrayList<>();
     private List<PodVolume> volumes = new ArrayList<PodVolume>();
     private WorkspaceVolume workspaceVolume;
@@ -49,7 +50,7 @@ public class PodTemplateStep extends Step implements Serializable {
     @DataBoundConstructor
     public PodTemplateStep(String label, String name) {
         this.label = label;
-        this.name = name == null ? "kubernetes" : name;
+        this.name = name == null ? "jenkins-slave" : name;
     }
 
     public String getLabel() {
@@ -58,6 +59,15 @@ public class PodTemplateStep extends Step implements Serializable {
 
     public String getName() {
         return name;
+    }
+
+    public String getNamespace() {
+        return namespace;
+    }
+
+    @DataBoundSetter
+    public void setNamespace(String namespace) {
+        this.namespace = namespace;
     }
 
     public String getCloud() {